+++ frontend/src/components/output/OutputArea.js
